LayoutTests/fast/css/internal-th-center-ua-only.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<style>
th {
    text-align: right;
    text-align: -internal-th-center;
}
</style>
<table>
  <tr>
    <th id="header"></th>
  </tr>
</table>

<script>
  test(function() {
    assert_false(CSS.supports('text-align', '-internal-th-center'));
  }, '"text-align" property does not support value "-internal-th-center".');

  test(function() {
    assert_equals(getComputedStyle(header).textAlign, 'right');
  }, '"text-align" property cannot be set to "-internal-th-center" by the author stylesheet.');
</script>
